Refatorar ChatArea extraindo Header e MessageList
