+++ src/components/chat_area.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useState } from 'react';
 import { useRef, useEffect } from 'react';
-import { MessageCircle, Send, Scale, User, Loader2 } from 'lucide-react';
+import { Send, Loader2 } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
+import ChatMsgHeader from '@/components/chat_msg_header';
+import ChatMsgList from '@/components/chat_msg_list';
 
 interface Message {
   id: number;
@@ -151,109 +151,10 @@ export default function ChatArea() {
   return (
     <div className="flex-1 flex flex-col bg-white" style={{ height: 'calc(100vh - 80px)' }}>
       {/* Header do Chat */}
-      <div className="flex flex-row items-center justify-between p-4 border-b border-neutral-200 bg-white">
-        <div className="flex items-center space-x-3">
-          <div className="flex h-8 w-8 items-center justify-center rounded-lg bg-gradient-to-br from-orange-500 to-orange-600">
-            <Scale className="h-4 w-4 text-white" />
-          </div>
-          <div>
-            <h3 className="text-lg font-semibold text-neutral-900">
-              Dante AI - Registro de Imóveis
-            </h3>
-            <div className="flex items-center space-x-2">
-              <div className="h-2 w-2 bg-green-500 rounded-full"></div>
-              <span className="text-xs text-neutral-500">Especialista Online</span>
-            </div>
-          </div>
-        </div>
-      </div>
+      <ChatMsgHeader />
 
       {/* Messages */}
-      <div className="flex-1 overflow-y-auto p-4 space-y-6 custom-scrollbar">
-        {messages.map((message) => (
-          <div
-            key={message.id}
-            className={`flex ${message.sender === 'user' ? 'justify-end' : 'justify-start'}`}
-          >
-            <div
-              className={`flex items-start space-x-3 max-w-[85%] ${
-                message.sender === 'user'
-                  ? 'flex-row-reverse space-x-reverse'
-                  : ''
-              }`}
-            >
-              {/* Avatar */}
-              <div
-                className={`flex h-8 w-8 items-center justify-center rounded-lg flex-shrink-0 ${
-                  message.sender === 'user'
-                    ? 'bg-neutral-900'
-                    : 'bg-gradient-to-br from-orange-500 to-orange-600'
-                }`}
-              >
-                {message.sender === 'user' ? (
-                  <User className="h-4 w-4 text-white" />
-                ) : (
-                  <Scale className="h-4 w-4 text-white" />
-                )}
-              </div>
-
-              {/* Conteúdo da mensagem */}
-              <div
-                className={`rounded-lg p-3 ${
-                  message.sender === 'user'
-                    ? 'bg-neutral-900 text-white'
-                    : 'bg-neutral-100 text-neutral-900'
-                }`}
-              >
-                {message.isLoading ? (
-                  <div className="flex items-center space-x-2">
-                    <Loader2 className="h-4 w-4 animate-spin" />
-                    <span className="text-sm">{message.loadingText || 'Processando...'}</span>
-                  </div>
-                ) : (
-                  <>
-                    {message.sender === 'bot' ? (
-                      <div className="prose prose-sm max-w-none prose-headings:text-neutral-900 prose-headings:font-semibold prose-h1:text-lg prose-h2:text-base prose-h3:text-sm prose-p:text-sm prose-p:leading-relaxed prose-strong:text-neutral-900 prose-strong:font-semibold prose-ul:text-sm prose-ol:text-sm prose-li:text-sm prose-blockquote:text-sm prose-blockquote:border-orange-300 prose-blockquote:bg-orange-50 prose-blockquote:px-3 prose-blockquote:py-2 prose-blockquote:rounded prose-code:text-xs prose-code:bg-neutral-200 prose-code:px-1 prose-code:rounded">
-                        <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                          {message.content}
-                        </ReactMarkdown>
-                      </div>
-                    ) : (
-                      <p className="text-sm leading-relaxed whitespace-pre-wrap">
-                        {message.content}
-                      </p>
-                    )}
-                    <div className="flex items-center justify-between mt-2">
-                      <span
-                        className={`text-xs ${
-                          message.sender === 'user'
-                            ? 'text-neutral-300'
-                            : 'text-neutral-500'
-                        }`}
-                      >
-                        {(() => {
-                          const date = message.timestamp;
-                          const day = date.getDate().toString().padStart(2, '0');
-                          const months = ['Jan', 'Fev', 'Mar', 'Abr', 'Mai', 'Jun', 
-                                         'Jul', 'Ago', 'Set', 'Out', 'Nov', 'Dez'];
-                          const month = months[date.getMonth()];
-                          const year = date.getFullYear().toString().slice(-2);
-                          const hours = date.getHours().toString().padStart(2, '0');
-                          const minutes = date.getMinutes().toString().padStart(2, '0');
-                          
-                          return `${day}/${month}/${year} - ${hours}:${minutes}`;
-                        })()}
-                      </span>
-                    </div>
-                  </>
-                )}
-              </div>
-            </div>
-          </div>
-        ))}
-        {/* Elemento invisível para scroll automático */}
-        <div ref={messagesEndRef} />
-      </div>
+      <ChatMsgList messages={messages} messagesEndRef={messagesEndRef} />
 
       {/* Input */}
       <div className="p-4 border-t border-neutral-200 bg-white">
